pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nanobanana"
version = "1.0.0"
description = "Google Gemini multimodal image generation agent for ACPX"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "Pillow>=10.0.0",
]

[project.scripts]
nanobanana = "nanobanana.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/nanobanana"]
